\name{AML2004}
\docType{data}
\alias{AML2004}
\title{Acute myelogenous leukemia dataset}
\description{
    Acute myelogenous leukemia dataset
}
\format{
    A list containing properties:
    \tabular{rlll}{
        Name          \tab Type           \tab Description\cr
        Gene          \tab data.frame     \tab mRNA expression data\cr
        Group         \tab data.frame     \tab Data frame indicating the cluster to which each sample is allocated\cr
    }
}
\source{\url{http://portals.broadinstitute.org/cgi-bin/cancer/publications/view/89}}
\references{
    Brunet, J. P., Tamayo, P., Golub, T. R., & Mesirov, J. P. (2004). Metagenes and molecular pattern discovery using matrix factorization. Proceedings of the national academy of sciences, 101(12), 4164-4169.
}
\keyword{datasets}